requirements.txt
# =============================================================================
# Merchant+ SaaS Financial Platform — Python Dependencies
# =============================================================================

# Core framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9

# Real-time WebSockets
channels[daphne]>=4.0
channels-redis>=4.1

# Filtering & Search
django-filter>=23.0

# CORS
django-cors-headers>=4.0

# Environment variables
python-decouple>=3.8

# Image handling (customer photos, ID docs, avatars)
Pillow>=10.0

# Production ASGI/WSGI servers
gunicorn>=21.0

# API Documentation — auto-generated OpenAPI/Swagger
drf-spectacular>=0.27

# Background tasks & scheduling
celery>=5.3
django-celery-beat>=2.5
redis>=5.0

# Webhook delivery
requests>=2.31

# Testing
pytest>=7.4
pytest-django>=4.7

# Two-Factor Authentication (TOTP)
pyotp>=2.9

# Encryption (customer photo encryption at rest)
cryptography>=42.0

# Utilities
pytz>=2023.3

# Static file serving (production without nginx)
whitenoise>=6.5
